requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
asttokens==3.0.0
attrs==25.4.0
beautifulsoup4==4.14.2
bitsandbytes==0.48.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
colorama==0.4.6
colorlog==6.10.1
comm==0.2.3
dataclasses-json==0.6.7
debugpy==1.8.17
decorator==5.2.1
dill==0.4.0
distro==1.9.0
docling==2.57.0
docling-core==2.49.0
docling-ibm-models==3.10.0
docling-parse==4.6.0
et_xmlfile==2.0.0
executing==2.2.1
Faker==37.11.0
filelock==3.20.0
filetype==1.2.0
frozenlist==1.8.0
fsspec==2025.9.0
ftfy==6.3.1
greenlet==3.2.4
grpcio==1.75.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface==0.0.1
huggingface-hub==0.35.3
idna==3.11
ipykernel==7.0.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
langchain==0.3.27
langchain-community==0.3.31
langchain-core==1.0.0
langchain-docling==1.1.0
langchain-huggingface==1.0.0
langchain-milvus==0.2.1
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langgraph==0.5.4
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.74
langsmith==0.4.37
latex2mathml==3.78.1
lxml==5.4.0
Markdown==3.9
markdown-it-py==4.0.0
marko==2.2.1
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.2.6
omegaconf==2.3.0
openai==2.5.0
opencv-python==4.12.0.88
openpyxl==3.1.5
orjson==3.11.3
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pillow==11.3.0
platformdirs==4.5.0
pluggy==1.6.0
polyfactory==2.22.2
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.0
pure_eval==0.2.3
pyclipper==1.3.0.post6
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
pylatexenc==2.10
pymilvus==2.6.2
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-pptx==1.0.2
pytz==2025.2
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
rapidocr==3.4.2
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.27.1
rtree==1.4.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
semchunk==2.2.2
sentence-transformers==3.4.1
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
stack-data==0.6.3
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.7.0+cu128
torchvision==0.22.0+cu128
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
typer==0.19.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
wcwidth==0.2.14
wheel==0.45.1
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
